requirements.txt
timm
Pillow 
numpy==1.*
albumentations
transformers[torch]
accelerate>=0.26.0
tensorboard
pycocotools